refactor worker liveness to executions gateway and enforce client fk
